requirements.txt
click~=8.0.4
pandas~=1.4.1
numpy~=1.22.2
xgboost~=1.5.2
scikit-learn~=1.0.2
mlflow~=1.23.1
PyYAML~=6.0